+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 * Add the method `DataElement.assertTokenName(String)` to simplify verification of an element when mapping to configuration objects.
 * Add the method `DataElement.assertTokenCount(int count)` to simplify verification of an element when mapping to configuration objects.
 * Add the method `DataElement.assertTokenCount(int min, int max)` to simplify verification of an element when mapping to configuration objects.
+* Add the methods `DataElement.getStringAt(int)`, `DataElement.getIntAt(int)` and `DataElement.getDoubleAt(int)` to simplify mapping elements to configuration objects.
 
 ### [v0.01](https://github.com/realityforge/zifnab/tree/v0.01) (2019-08-06)
 [Full Changelog](https://github.com/realityforge/zifnab/compare/b24bbdea2237c119e17341e5597c42b21b76a9c9...v0.01)